Made show page work with live API and Error Reporting.
Picture needed to be passed into the controller. This was defined in the services.
Error reporting was added, to pass it to view, needed to make a scope.
ng-show is conditional. Error will display if an error exists.
ID in index (to allow pass through to show) is defined with picture.id.